Clean Ruby files
